# Update perf script