Now this can run with any board size and with either Python or C++ implementations.

Some numbers from my M2 Macbook:

```
$ poetry run python -m boggle.perf --size 33 100000 --random_seed 808813
Generating 100000 3x3 boards...
Scoring boards...
total_score=1255148
0.15s, 660266.61 bds/sec

$ poetry run python -m boggle.perf --size 33 100000 --random_seed 808813 --python
Generating 100000 3x3 boards...
Scoring boards...
total_score=1255148
3.36s, 29765.27 bds/sec

$ poetry run python -m boggle.perf --size 34 1000000 --random_seed 808813
Generating 1000000 3x4 boards...
Scoring boards...
total_score=22763453
2.59s, 385896.28 bds/sec

poetry run python -m boggle.perf --size 44 1000000 --random_seed 808813
Generating 1000000 4x4 boards...
Scoring boards...
total_score=41134010
4.55s, 219755.78 bds/sec
```